+++ Extension/src/background/api/filtering-log.ts
@@ -19,9 +19,6 @@
  */
 import { type Tabs } from 'webextension-polyfill';
 
-import { RuleGenerator } from '@adguard/agtree/generator';
-import { RULE_INDEX_NONE } from '@adguard/tsurlfilter';
-
 import {
     BACKGROUND_TAB_ID,
     type ContentType,
@@ -29,9 +26,6 @@ import {
     isExtensionUrl,
     type StealthActionEvent,
     getDomain,
-    getRuleSourceText,
-    getRuleSourceIndex,
-    type PreprocessedFilterList,
     type DeclarativeRuleInfo,
 } from 'tswebextension';
 
@@ -42,12 +36,13 @@ import { engine } from '../engine';
 import { settingsStorage } from '../storages/settings';
 import { SettingOption } from '../schema/settings/enum';
 import { TabsApi } from '../../common/api/extension/tabs';
-import { AntiBannerFiltersId, NotifierType } from '../../common/constants';
-import { FiltersStoragesAdapter } from '../storages/filters-adapter';
+import { NotifierType } from '../../common/constants';
 
 export type FilteringEventRuleData = {
     filterId: number;
     ruleIndex: number;
+    appliedRuleText: string;
+    originalRuleText?: string;
     isImportant?: boolean;
     documentLevelRule?: boolean;
     isStealthModeRule?: boolean;
@@ -159,25 +154,11 @@ interface RuleText {
 }
 
 /**
- * Cached filter data.
- */
-type CachedFilterData = Pick<PreprocessedFilterList, 'rawFilterList' | 'conversionMap' | 'sourceMap'>;
-
-/**
  * The filtering log collects all available information about requests
  * and the rules applied to them.
  */
 export class FilteringLogApi {
     private static readonly REQUESTS_SIZE_PER_TAB = 1000;
-
-    /**
-     * Filter lists that are generated dynamically by TSWebExtension.
-     * We don't store them in the storage, so we need to get rule AST nodes and generate rule text manually.
-     */
-    private static readonly DYNAMIC_FILTER_LISTS = new Set([
-        AntiBannerFiltersId.StealthModeFilterId,
-        AntiBannerFiltersId.AllowlistFilterId,
-    ]);
 
     private openedFilteringLogsPages = 0;
 
@@ -190,272 +171,6 @@ export class FilteringLogApi {
             domain: null,
         }],
     ]);
-
-    /**
-     * Cache for filters data.
-     *
-     * The key is the filter list id and the value is the filter data.
-     * This cache is used to avoid unnecessary requests to the storage while filtering log is opened.
-     * After closing the filtering log, the cache is purged to free up memory.
-     */
-    private filtersCache = new Map<number, CachedFilterData>();
-
-    /**
-     * Map of filter sync attempts. The key is the filter list id and the value is the last attempt time.
-     * In some rare edge cases (but maybe never), the filters cache may be become outdated,
-     * so we need to sync the filter again.
-     * We only try this sync if we encounter an error while getting rule text.
-     */
-    private filterSyncAttempts = new Map<number, number>();
-
-    /**
-     * Interval for sync attempts in milliseconds.
-     * We should try to sync the filter only once per 10 seconds.
-     */
-    private static readonly SYNC_ATTEMPTS_INTERVAL_MS = 10000;
-
-    /**
-     * Flag to invert allowlist mode.
-     */
-    private allowlistInverted = false;
-
-    /**
-     * Purges filters cache.
-     *
-     * @param filterIds Filter ids to remove from cache. If not provided, the whole cache will be purged.
-     */
-    private purgeFiltersCache(filterIds?: number[]): void {
-        if (filterIds) {
-            filterIds.forEach((filterId) => this.filtersCache.delete(filterId));
-
-            logger.debug(`[ext.FilteringLogApi.purgeFiltersCache]: filtering log filters cache purged for filter ids: ${filterIds.join(', ')}`);
-            return;
-        }
-
-        this.filtersCache.clear();
-
-        logger.debug('[ext.FilteringLogApi.purgeFiltersCache]: filtering log filters cache purged.');
-    }
-
-    /**
-     * Purges filter sync attempts.
-     *
-     * @param filterIds Filter ids to remove from cache. If not provided, the whole cache will be purged.
-     */
-    private purgeFilterSyncAttempts(filterIds?: number[]): void {
-        if (filterIds) {
-            filterIds.forEach((filterId) => this.filterSyncAttempts.delete(filterId));
-            return;
-        }
-
-        this.filterSyncAttempts.clear();
-    }
-
-    /**
-     * Called when the engine is updated.
-     *
-     * @param allowlistInverted Whether allowlist mode is inverted.
-     */
-    public onEngineUpdated(allowlistInverted: boolean): void {
-        this.allowlistInverted = allowlistInverted;
-
-        if (!this.isOpen()) {
-            return;
-        }
-
-        // We need to purge the filters cache, because the filters data may be changed after the engine update
-        this.purgeFiltersCache();
-        this.purgeFilterSyncAttempts();
-    }
-
-    /**
-     * Helper method to get filter data.
-     * It handles a cache internally to speed up requests for the same filter next time.
-     *
-     * @param filterId Filter id.
-     *
-     * @returns Filter data or undefined if the filter is not found.
-     */
-    private async getFilterData(filterId: number): Promise<CachedFilterData | undefined> {
-        if (this.filtersCache.has(filterId)) {
-            return this.filtersCache.get(filterId);
-        }
-
-        try {
-            // eslint-disable-next-line prefer-const
-            let [rawFilterList, conversionMap, sourceMap] = await Promise.all([
-                FiltersStoragesAdapter.getRawFilterList(filterId),
-                FiltersStoragesAdapter.getConversionMap(filterId),
-                FiltersStoragesAdapter.getSourceMap(filterId),
-            ]);
-
-            // Raw filter list and source map are required to get rule texts
-            if (!rawFilterList || !sourceMap) {
-                return undefined;
-            }
-
-            if (!conversionMap) {
-                conversionMap = {};
-            }
-
-            const filterData: CachedFilterData = {
-                rawFilterList,
-                conversionMap,
-                sourceMap,
-            };
-
-            this.filtersCache.set(filterId, filterData);
-
-            return filterData;
-        } catch (e) {
-            logger.error(`[ext.FilteringLogApi.getFilterData]: failed to get filter data for filter id ${filterId}:`, e);
-        }
-
-        return undefined;
-    }
-
-    /**
-     * Tries to sync the cached filter data with the storage.
-     *
-     * @param filterId Filter id.
-     *
-     * @returns Returns false if the filter was synced within the last
-     * {@link FilteringLogApi.SYNC_ATTEMPTS_INTERVAL_MS} milliseconds.
-     * Otherwise, returns true.
-     */
-    private attemptToSyncFilter(filterId: number): boolean {
-        const lastAttemptTime = this.filterSyncAttempts.get(filterId) || 0;
-
-        if (Date.now() - lastAttemptTime < FilteringLogApi.SYNC_ATTEMPTS_INTERVAL_MS) {
-            return false;
-        }
-
-        // Delete filter from the cache to force a new request next time
-        this.purgeFiltersCache([filterId]);
-
-        this.filterSyncAttempts.set(filterId, Date.now());
-
-        return true;
-    }
-
-    /**
-     * Gets rule text for the specified filter id and rule index.
-     * If the rule is not found, returns null.
-     * It handles a cache internally to speed up requests for the same filter next time.
-     *
-     * @param filterId Filter id.
-     * @param ruleIndex Rule index.
-     *
-     * @returns Rule text or null if the rule is not found.
-     */
-    public async getRuleText(filterId: number, ruleIndex: number): Promise<RuleText | null> {
-        // Special case: when inverted allowlist mode is enabled, we generate rules dynamically without rule index,
-        // and we don't need to show them, so we return null.
-        // https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2883
-        if (
-            this.allowlistInverted
-            && filterId === AntiBannerFiltersId.AllowlistFilterId
-            && ruleIndex === RULE_INDEX_NONE
-        ) {
-            return null;
-        }
-
-        // Note: Stealth and Allowlist are special filters, they're generated dynamically by TSWebExtension internally.
-        // We don't store them in the storage, so we need to get rule AST nodes and generate rule text manually.
-        if (FilteringLogApi.DYNAMIC_FILTER_LISTS.has(filterId)) {
-            const ruleNode = engine.api.retrieveRuleNode(filterId, ruleIndex);
-
-            // The following error messages should not be displayed during normal operation,
-            // but we handle them just in case, and to provide type safety
-            if (!ruleNode) {
-                logger.error(`[ext.FilteringLogApi.getRuleText]: failed to get rule node for filter id ${filterId} and rule index ${ruleIndex}`);
-                return null;
-            }
-
-            const ruleText = RuleGenerator.generate(ruleNode);
-
-            if (!ruleText) {
-                logger.error(`[ext.FilteringLogApi.getRuleText]: failed to get rule text for filter id ${filterId} and rule index ${ruleIndex}`);
-                return null;
-            }
-
-            return {
-                appliedRuleText: ruleText,
-            };
-        }
-
-        const filterData = await this.getFilterData(filterId);
-
-        if (!filterData) {
-            logger.error(`[ext.FilteringLogApi.getRuleText]: failed to get filter data for filter id ${filterId}`);
-            return null;
-        }
-
-        const { rawFilterList, conversionMap, sourceMap } = filterData;
-
-        // Get line start index in the source file by rule start index in the byte array
-        const lineStartIndex = getRuleSourceIndex(ruleIndex, sourceMap);
-
-        if (lineStartIndex === -1) {
-            let baseMessage = `[ext.FilteringLogApi.getRuleText]: failed to get line start index for filter id ${filterId} and rule index ${ruleIndex}`;
-
-            const ruleNode = engine.api.retrieveRuleNode(filterId, ruleIndex);
-
-            // Note: during normal operation, ruleNode should not be null,
-            // but we handle this case just in case, and to provide type safety
-            if (ruleNode) {
-                const generatedRuleText = RuleGenerator.generate(ruleNode);
-                baseMessage += `, generated rule text: ${generatedRuleText}`;
-            }
-
-            // If the rule is not found, try to sync the filter and try again
-            if (this.attemptToSyncFilter(filterId)) {
-                // eslint-disable-next-line @adguard/logger-context/require-logger-context
-                logger.warn(`${baseMessage}, trying to sync the filter`);
-                return this.getRuleText(filterId, ruleIndex);
-            }
-
-            // eslint-disable-next-line @adguard/logger-context/require-logger-context
-            logger.error(baseMessage);
-            return null;
-        }
-
-        const appliedRuleText = getRuleSourceText(lineStartIndex, rawFilterList);
-
-        if (!appliedRuleText) {
-            let baseMessage = `[ext.FilteringLogApi.getRuleText]: failed to get rule text for filter id ${filterId} and rule index ${ruleIndex}`;
-
-            const ruleNode = engine.api.retrieveRuleNode(filterId, ruleIndex);
-
-            // Note: during normal operation, ruleNode should not be null,
-            // but we handle this case just in case, and to provide type safety
-            if (ruleNode) {
-                const generatedRuleText = RuleGenerator.generate(ruleNode);
-                baseMessage += `, generated rule text: ${generatedRuleText}`;
-            }
-
-            // If the rule is not found, try to sync the filter and try again
-            if (this.attemptToSyncFilter(filterId)) {
-                // eslint-disable-next-line @adguard/logger-context/require-logger-context
-                logger.warn(`${baseMessage}, trying to sync the filter`);
-                return this.getRuleText(filterId, ruleIndex);
-            }
-
-            // eslint-disable-next-line @adguard/logger-context/require-logger-context
-            logger.error(baseMessage);
-            return null;
-        }
-
-        const result: RuleText = {
-            appliedRuleText,
-        };
-
-        if (conversionMap && conversionMap[lineStartIndex]) {
-            result.originalRuleText = conversionMap[lineStartIndex];
-        }
-
-        return result;
-    }
 
     /**
      * Checks if filtering log page is opened.
@@ -509,10 +224,6 @@ export class FilteringLogApi {
     public onCloseFilteringLogPage(): void {
         this.openedFilteringLogsPages = Math.max(this.openedFilteringLogsPages - 1, 0);
         if (this.openedFilteringLogsPages === 0) {
-            // Purge filters cache to free up memory
-            this.purgeFiltersCache();
-            this.purgeFilterSyncAttempts();
-
             // Clear events
             this.tabsInfoMap.forEach((tabInfo) => {
                 tabInfo.filteringEvents = [];
@@ -701,11 +412,6 @@ export class FilteringLogApi {
             return;
         }
 
-        // Get rule text based on filter id and rule index
-        if (data.requestRule) {
-            data.requestRule = await this.applyRuleTextToRuleData(data.requestRule);
-        }
-
         tabInfo.filteringEvents.push(data);
 
         const shouldRemoveOldestEvent = !FilteringLogApi.isPreserveLogEnabled()
@@ -742,45 +448,7 @@ export class FilteringLogApi {
             return;
         }
 
-        if (data.requestRule && !event.requestRule?.appliedRuleText) {
-            data.requestRule = await this.applyRuleTextToRuleData(data.requestRule);
-        }
-
-        if (data.replaceRules) {
-            data.replaceRules = await this.applyRuleTextToRuleDataArray(data.replaceRules);
-        }
-
-        if (data.stealthAllowlistRules) {
-            data.stealthAllowlistRules = await this.applyRuleTextToRuleDataArray(data.stealthAllowlistRules);
-        }
-
         event = Object.assign(event, data);
-    }
-
-    /**
-     * Retrieves and applies rule text to a single rule data object.
-     *
-     * @param rule Rule data object containing filterId and ruleIndex.
-     *
-     * @returns Rule object with rule text applied.
-     */
-    private async applyRuleTextToRuleData(rule: FilteringEventRuleData): Promise<FilteringEventRuleData> {
-        const { filterId, ruleIndex } = rule;
-        const ruleTextData = await this.getRuleText(filterId, ruleIndex);
-        return ruleTextData ? Object.assign(rule, ruleTextData) : rule;
-    }
-
-    /**
-     * Retrieves and applies rule text to an array of rule data objects.
-     *
-     * @param rules Array of rule data objects containing filterId and ruleIndex.
-     *
-     * @returns Array of rule objects with rule text applied.
-     */
-    private async applyRuleTextToRuleDataArray(rules: FilteringEventRuleData[]): Promise<FilteringEventRuleData[]> {
-        return Promise.all(
-            rules.map(async (rule) => this.applyRuleTextToRuleData(rule)),
-        );
     }
 
     /**