+++ Extension/src/background/api/ui/icons/icons-mv3.ts
@@ -18,8 +18,7 @@
  * along with AdGuard Browser Extension. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { RulesLimitsService } from 'rules-limits-service';
-
+import { RulesLimitsService } from '../../../services/rules-limits/rules-limits-service-mv3';
 import { ExtensionUpdateService } from '../../../services/extension-update/extension-update-service-mv3';
 import {
     appContext,