+++ Extension/src/background/engine/engine-mv3.ts
@@ -29,6 +29,7 @@ import {
     type MessageHandler,
     type Message as EngineMessage,
     type ConfigurationResult,
+    FilterList,
 } from '@adguard/tswebextension/mv3';
 
 import { logger } from '../../common/logger';
@@ -47,7 +48,7 @@ import {
 } from '../api';
 import { RulesLimitsService, rulesLimitsService } from '../services/rules-limits/rules-limits-service-mv3';
 import { UserRulesService } from '../services/userrules';
-import { emptyPreprocessedFilterList, NotifierType } from '../../common/constants';
+import { NotifierType } from '../../common/constants';
 import { SettingOption } from '../schema/settings/enum';
 import { localScriptRules } from '../../../filters/chromium-mv3/local_script_rules';
 import { FiltersStorage } from '../storages/filters';
@@ -159,7 +160,6 @@ export class Engine implements TsWebExtensionEngine {
          *   - Default filters (IDs: 2, 10) are pending enablement.
          */
         await iconsApi.update();
-        filteringLogApi.onEngineUpdated(configuration.settings.allowlistInverted);
     }
 
     /**
@@ -197,7 +197,6 @@ export class Engine implements TsWebExtensionEngine {
         }
         // Updates extension icon state to reflect current filtering status.
         await iconsApi.update();
-        filteringLogApi.onEngineUpdated(configuration.settings.allowlistInverted);
     }
 
     /**
@@ -221,17 +220,18 @@ export class Engine implements TsWebExtensionEngine {
             }
         }
 
-        const userrules: Configuration['userrules'] = {
-            ...emptyPreprocessedFilterList,
-            /**
-             * User rules are always trusted.
-             */
-            trusted: true,
-        };
+        let userRulesFilter: FilterList;
 
         if (UserRulesApi.isEnabled()) {
-            Object.assign(userrules, await UserRulesApi.getUserRules());
+            userRulesFilter = await UserRulesApi.getUserRules();
+        } else {
+            userRulesFilter = FilterList.createEmpty();
         }
+
+        const userrules: Configuration['userrules'] = {
+            content: userRulesFilter.getContent(),
+            conversionData: userRulesFilter.getConversionData(),
+        };
 
         let customFilters: Configuration['customFilters'] = [];
 
@@ -287,12 +287,17 @@ export class Engine implements TsWebExtensionEngine {
                 .filter((f) => CustomFilterApi.isCustomFilterMetadata(f));
 
             customFilters = await Promise.all(customFiltersWithMetadata.map(async ({ filterId, trusted }) => {
-                const preprocessedFilterList = await FiltersStorage.get(filterId);
+                let filterList = await FiltersStorage.get(filterId);
+
+                if (!filterList) {
+                    filterList = FilterList.createEmpty();
+                }
 
                 return {
                     filterId,
+                    content: filterList.getContent(),
+                    conversionData: filterList.getConversionData(),
                     trusted,
-                    ...(preprocessedFilterList || emptyPreprocessedFilterList),
                 };
             }));
         }