+++ Extension/src/background/services/filters/filters-service-mv3.ts
@@ -17,9 +17,6 @@
  * You should have received a copy of the GNU General Public License
  * along with AdGuard Browser Extension. If not, see <http://www.gnu.org/licenses/>.
  */
-
-import { RulesLimitsService } from 'rules-limits-service';
-
 import {
     type AddAndEnableFilterMessage,
     type DisableFilterMessage,
@@ -30,6 +27,7 @@ import {
     type SetConsentedFiltersMessage,
 } from '../../../common/messages';
 import { logger } from '../../../common/logger';
+import { RulesLimitsService } from '../rules-limits/rules-limits-service-mv3';
 import { messageHandler } from '../../message-handler';
 import { engine } from '../../engine';
 import {