+++ Extension/src/pages/options/components/Filters/Filters.tsx
@@ -470,16 +470,16 @@ const Filters = observer(() => {
             >
                 <div className="title__inner">
                     <div className="title title--back-btn">
+                        <div className="button setting__back">
+                            <Icon
+                                id="#arrow-left"
+                                className="icon--24"
+                                aria-hidden="true"
+                            />
+                        </div>
                         <span id={titleId}>{selectedGroup.groupName}</span>
                     </div>
                     {description && <div className="title__desc title__desc--back">{description}</div>}
-                </div>
-                <div className="button setting__back">
-                    <Icon
-                        id="#arrow-left"
-                        className="icon--24"
-                        aria-hidden="true"
-                    />
                 </div>
             </button>
         );